feat: create schema of collector service
