+++ apps/web/src/schema/inventory/collector/api-verbs/delete.ts
@@ -1,0 +1,4 @@
+export interface CollectorDeleteParameters {
+    collector_id: string;
+    workspace_id?: string;
+}
