+++ apps/web/src/schema/inventory/collector/api-verbs/get.ts
@@ -1,0 +1,4 @@
+export interface CollectorGetParameters {
+    collector_id: string;
+    workspace_id?: string;
+}
