+++ apps/web/src/schema/inventory/collector/api-verbs/stat.ts
@@ -1,0 +1,3 @@
+export interface CollectorStatParameters {
+    query: any;
+}
