+++ apps/web/src/schema/inventory/collector/api-verbs/update-plugin.ts
@@ -1,0 +1,8 @@
+import type { CollectorOptions, UpgradeMode } from '@/schema/inventory/collector/type';
+
+export interface CollectorUpdatePluginParameters {
+    collector_id: string;
+    version?: string;
+    options?: CollectorOptions; // backend api will replace whole schedule object
+    upgrade_mode?: UpgradeMode;
+}
