+++ apps/web/src/schema/inventory/collector/constant.ts
@@ -1,0 +1,10 @@
+// CollectorPluginModel
+export const UPGRADE_MODE = {
+    AUTO: 'AUTO',
+    MANUAL: 'MANUAL',
+} as const;
+// CollectorModel
+export const STATE = {
+    ENABLED: 'ENABLED',
+    DISABLED: 'DISABLED',
+} as const;
