+++ apps/web/src/services/asset-inventory/stores/collector-page-store.ts
@@ -4,7 +4,8 @@ import type { ConsoleFilter } from '@cloudforet/core-lib/query/type';
 import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 import type { Query } from '@cloudforet/core-lib/space-connector/type';
 
-import type { CollectorModel, Schedule } from '@/schema/inventory/collector/model';
+import type { CollectorModel } from '@/schema/inventory/collector/model';
+import type { Schedule } from '@/schema/inventory/collector/type';
 import type { JobModel } from '@/schema/inventory/job/model';
 import { i18n } from '@/translations';
